Move useDropPhotoFiles to hooks
